chore: update oauth

+++ src/modules/credentials/credentials.service.ts
@@ -1,5 +1,5 @@
 import {
-  Injectable, Inject, forwardRef, ForbiddenException, UnauthorizedException,
+  Injectable, Inject, forwardRef, ForbiddenException, UnauthorizedException, BadGatewayException,
 } from '@nestjs/common';
 import { Repository } from 'typeorm';
 import { InjectRepository } from '@nestjs/typeorm';
@@ -43,13 +43,39 @@ export class CredentialsService {
     if (!strData) {
       throw new UnauthorizedException('code credentials are invalid');
     }
-    // const { data, type } = JSON.parse(strData);
+    const { data, type } = JSON.parse(strData);
+    const isAuthorize = await Promise.all([
+      this.credentialsRepository.createQueryBuilder('cr')
+        .where('cr.userId=:id AND cr.type=:type', { id: user.id, type })
+        .getOne(),
+      this.credentialsRepository.createQueryBuilder('cr')
+        .where('cr.id=:id', { id: `${type}_${data.id}` })
+        .getOne(),
+    ]);
+    if (isAuthorize[0]) {
+      throw new UnauthorizedException(`authorized ${type}`);
+    }
+    if (isAuthorize[1]) {
+      throw new UnauthorizedException('github account authorized');
+    }
+    await this.create({
+      id: `${type}_${data.id}`,
+      type,
+      user,
+      info: data,
+    });
   }
 
   public async revoke(user: UserEntity, id: string) {
     const data = await this.getInfo(id);
     if (!data || data.user.id !== user.id) {
       throw new ForbiddenException();
+    }
+    const count = await this.credentialsRepository.createQueryBuilder('cr')
+      .where('cr.userId=:id', { id: user.id })
+      .getCount();
+    if (count === 1 && !user.isPassword) {
+      throw new BadGatewayException('reserved login type');
     }
     await this.credentialsRepository.createQueryBuilder()
       .delete()
